requirements.txt
requests
websocket-client
# clipboard
pyperclip